requirements.txt
groq
RealtimeSTT
pygame
edge_tts
asyncio
pillow
bs4